chore(package.json): update author and bugs fields with correct information
The author field in the package.json file was updated to include the name, email, and URL of the author. This change was made to provide accurate information about the author of the project.

The bugs field in the package.json file was also updated to include the URL and email address for reporting issues. This change was made to ensure that users can easily report any bugs they encounter while using the project.

No functional changes were made to the code or dependencies in this commit.