feat: add .prettierrc file with formatting configurations
The .prettierrc file is added to the project with the following formatting configurations:
- tabWidth: 4
- useTabs: false
- singleQuote: true
- semi: true
- endOfLine: lf
- bracketSpacing: true
- printWidth: 120
- trailingComma: all

This will ensure consistent code formatting across the project.

chore(package.json): update dependencies

The axios package is added as a dependency with version "^1.6.5" to enable making HTTP requests.
The prettier package is added as a dev dependency with version "^3.2.4" to support code formatting.

This update improves the functionality and maintainability of the project.